ADD GIT-FILE-NAME
Since the old source tarbell is cached, when leaving the old hash
when updating/building, this can cause the old version to rebuild instead of
the new source if you only change the 'version' in the package definition and
an explicit git-file-name is not used.

This behavior is a fallback to keep packages building if the upstream
is down or vanishes.

Fix this by making the file-name explicit to the version.